#RAVI

+++ Kaushal_Darpan.Models/Revaluation/ITIStudentRevaluationDataModel.cs
@@ -8,7 +8,7 @@ namespace Kaushal_Darpan.Models.RevaluationDataModel
     {
 
         public string DOB { get; set; }
-        public int? RollNo { get; set; }
+        public Int64? RollNo { get; set; }
         public int DepartmentID { get; set; }
         public int CourseTypeIDs { get; set; }
